# Speed up and profile error analysis