--- a/src/main/scala/wordbots/package.scala
+++ b/src/main/scala/wordbots/package.scala
@@ -7,4 +7,16 @@ package object wordbots {
   def parse(cmd: String): Unit = {
     Parser.main(Array(cmd))
   }
+
+  val simpleTokenizer: String => IndexedSeq[String] = { str: String =>
+    str
+      .trim
+      .toLowerCase
+      .replaceAll("[\u202F\u00A0]", " ")  // treat special space characters as spaces
+      .replaceAllLiterally("\' ", " \' ")  // add spaces before and after <'> to make it a separate token
+      .replaceAllLiterally("\'s", " \'s ")  // add spaces before and after <'s> to make it a separate token
+      .replaceAll("""["<>]""", " $0 ")  // add spaces before and after { " < > }, to make them separate tokens
+      .split("""\s+|[.?!,()]""")  // tokenize by splitting on spaces and punctuation
+      .filter("" !=)  // ignore empty tokens
+  }
 }