… src/main/scala/wordbots/Lexicon.scala