A variety of improvements with the overall goal of speeding up error analysis, currently the bottleneck of the parsing pipeline:
* track both the amount of time and the # successful candidates / # tried candidates for syntactic and semantic parses during error analysis
* cache candidate parse results
* truncate the list of successful candidate parses to 10 and stop trying them once we have enough
* order terms based on frequency in current set of successful parses (exported from Firebase) so more frequently-used terms are tried first

As an example, for two tricky cases, error analysis times previously were:
```
> parse("After the end of your turn draw 4 cards")
Spent 119 ms syntactic parsing (23/123) and 1793 ms semantic parsing (1/2090)

> parse("Activate: gain +1 health and +1 attack")
Spent 71 ms syntactic parsing (11/97) and 3285 ms semantic parsing (60/445)
```

And are now:
```
> parse("After the end of your turn draw 4 cards")
Spent 130 ms syntactic parsing (23/123) and 1014 ms semantic parsing (1/1600)

> parse("Activate: gain +1 health and +1 attack")
Spent 15 ms syntactic parsing (5/19) and 248 ms semantic parsing (11/28)
```